Use Java language extension to get Java Home
This means the extension and its checking will always match what the language extension requires.

There is one downside with this, in that this does require the redhat extension to fully start up. This can delay loading of this extension, as that requires model syncing, but maybe I can work with that extension to fix that issue.